separate classification and CV code
