+++ 02_classify_stratified/run_purity_prediction.py
@@ -16,7 +16,7 @@ from mpmp.exceptions import (
     NoTrainSamplesError,
     OneClassError,
 )
-from mpmp.utilities.classify_utilities import run_cv_stratified
+from mpmp.prediction.classification import run_cv_stratified
 import mpmp.utilities.data_utilities as du
 import mpmp.utilities.file_utilities as fu
 from mpmp.utilities.tcga_utilities import get_overlap_data_types
