+++ mpmp/prediction/classification.py
@@ -1,7 +1,7 @@
 """
 Functions for training classifiers on TCGA data.
 
-Many of these functions are adapted from:
+Some of these functions are adapted from:
 https://github.com/greenelab/BioBombe/blob/master/9.tcga-classify/scripts/tcga_util.py
 """
 import warnings
@@ -17,12 +17,12 @@ from sklearn.metrics import (
     average_precision_score
 )
 from sklearn.model_selection import (
-    StratifiedKFold,
     cross_val_predict,
     GridSearchCV,
 )
 
 import mpmp.config as cfg
+import mpmp.prediction.cross_validation as cv
 import mpmp.utilities.tcga_utilities as tu
 from mpmp.exceptions import (
     NoTrainSamplesError,
@@ -77,7 +77,7 @@ def run_cv_stratified(data_model,
                 # can ignore that warning.
                 warnings.filterwarnings('ignore',
                                         message='The least populated class in y')
-                X_train_raw_df, X_test_raw_df, _ = split_stratified(
+                X_train_raw_df, X_test_raw_df, _ = cv.split_stratified(
                    data_model.X_df, sample_info, num_folds=num_folds,
                    fold_no=fold_no, seed=data_model.seed)
         except ValueError:
@@ -425,58 +425,4 @@ def summarize_results(results,
 
     return metrics_out_, roc_df_, pr_df_
 
-
-def split_stratified(data_df,
-                     sample_info_df,
-                     num_folds=4,
-                     fold_no=1,
-                     seed=cfg.default_seed):
-    """Split expression data into train and test sets.
-
-    The train and test sets will both contain data from all cancer types,
-    in roughly equal proportions.
-
-    Arguments
-    ---------
-    data_df (pd.DataFrame): samples x features dataframe
-    sample_info_df (pd.DataFrame): maps samples to cancer types
-    num_folds (int): number of cross-validation folds
-    fold_no (int): cross-validation fold to hold out
-    seed (int): seed for deterministic splits
-
-    Returns
-    -------
-    train_df (pd.DataFrame): samples x features train data
-    test_df (pd.DataFrame): samples x features test data
-    """
-    # subset sample info to samples in pre-filtered expression data
-    sample_info_df = sample_info_df.reindex(data_df.index)
-
-    # generate id for stratification
-    # this is a concatenation of cancer type and sample/tumor type, since we want
-    # to stratify by both
-    sample_info_df = sample_info_df.assign(
-        id_for_stratification = sample_info_df.cancer_type.str.cat(
-                                                sample_info_df.sample_type)
-    )
-    # recode stratification id if they are singletons or near-singletons,
-    # since these won't work with StratifiedKFold
-    stratify_counts = sample_info_df.id_for_stratification.value_counts().to_dict()
-    sample_info_df = sample_info_df.assign(
-        stratify_samples_count = sample_info_df.id_for_stratification
-    )
-    sample_info_df.stratify_samples_count = sample_info_df.stratify_samples_count.replace(
-        stratify_counts)
-    sample_info_df.loc[
-        sample_info_df.stratify_samples_count < num_folds, 'id_for_stratification'
-    ] = 'other'
-
-    # now do stratified CV splitting and return the desired fold
-    kf = StratifiedKFold(n_splits=num_folds, shuffle=True, random_state=seed)
-    for fold, (train_ixs, test_ixs) in enumerate(
-            kf.split(data_df, sample_info_df.id_for_stratification)):
-        if fold == fold_no:
-            train_df = data_df.iloc[train_ixs]
-            test_df = data_df.iloc[test_ixs]
-    return train_df, test_df, sample_info_df
