+++ mpmp/prediction/cross_validation.py
@@ -1,0 +1,63 @@
+"""
+Functions to split data for cross-validation.
+
+"""
+import pandas as pd
+from sklearn.model_selection import StratifiedKFold
+
+import mpmp.config as cfg
+
+def split_stratified(data_df,
+                     sample_info_df,
+                     num_folds=4,
+                     fold_no=1,
+                     seed=cfg.default_seed):
+    """Split expression data into train and test sets.
+
+    The train and test sets will both contain data from all cancer types,
+    in roughly equal proportions.
+
+    Arguments
+    ---------
+    data_df (pd.DataFrame): samples x features dataframe
+    sample_info_df (pd.DataFrame): maps samples to cancer types
+    num_folds (int): number of cross-validation folds
+    fold_no (int): cross-validation fold to hold out
+    seed (int): seed for deterministic splits
+
+    Returns
+    -------
+    train_df (pd.DataFrame): samples x features train data
+    test_df (pd.DataFrame): samples x features test data
+    """
+    # subset sample info to samples in pre-filtered expression data
+    sample_info_df = sample_info_df.reindex(data_df.index)
+
+    # generate id for stratification
+    # this is a concatenation of cancer type and sample/tumor type, since we want
+    # to stratify by both
+    sample_info_df = sample_info_df.assign(
+        id_for_stratification = sample_info_df.cancer_type.str.cat(
+                                                sample_info_df.sample_type)
+    )
+    # recode stratification id if they are singletons or near-singletons,
+    # since these won't work with StratifiedKFold
+    stratify_counts = sample_info_df.id_for_stratification.value_counts().to_dict()
+    sample_info_df = sample_info_df.assign(
+        stratify_samples_count = sample_info_df.id_for_stratification
+    )
+    sample_info_df.stratify_samples_count = sample_info_df.stratify_samples_count.replace(
+        stratify_counts)
+    sample_info_df.loc[
+        sample_info_df.stratify_samples_count < num_folds, 'id_for_stratification'
+    ] = 'other'
+
+    # now do stratified CV splitting and return the desired fold
+    kf = StratifiedKFold(n_splits=num_folds, shuffle=True, random_state=seed)
+    for fold, (train_ixs, test_ixs) in enumerate(
+            kf.split(data_df, sample_info_df.id_for_stratification)):
+        if fold == fold_no:
+            train_df = data_df.iloc[train_ixs]
+            test_df = data_df.iloc[test_ixs]
+    return train_df, test_df, sample_info_df
+
